Added Regression Refactoring of classes into single model base class
